labs/ai_assisted_local_llm/pyproject.toml
[project]
name = "ai-assisted-llm-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.40.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "pandas>=2.2.0",
    "piper-tts>=1.4.1",
    "plotly>=5.24.0",
    "pydantic>=2.12.5",
    "python-docx>=1.1.0",
    "python-dotenv>=1.1.0",
    "redis>=7.2.0",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.15.1",
]
